Changes the return type when coercing attributes.
Previously the value was extracted by the attr handler, returning both
the name and the coerced value.

Now the rec caller will extract the name and just pass the value for
coercion.

This makes the coercing logic more uniform as the attr's are now
treated pretty much like any other type of entity.